+++ app/src/main/java/org/linphone/activities/assistant/fragments/PhoneAccountCreationFragment.kt
@@ -62,6 +62,10 @@ class PhoneAccountCreationFragment :
             countryPickerFragment.show(childFragmentManager, "CountryPicker")
         }
 
+        viewModel.prefix.observe(viewLifecycleOwner) { internationalPrefix ->
+            viewModel.getCountryNameFromPrefix(internationalPrefix)
+        }
+
         viewModel.goToSmsValidationEvent.observe(
             viewLifecycleOwner
         ) {