+++ app/src/main/java/org/linphone/activities/main/contact/viewmodels/ContactsListViewModel.kt
@@ -119,17 +119,17 @@ class ContactsListViewModel : ViewModel() {
         previousFilter = filterValue
 
         val domain = if (sipContactsSelected.value == true) coreContext.core.defaultAccount?.params?.domain ?: "" else ""
-        val filter = MagicSearch.Source.Friends.toInt() or MagicSearch.Source.LdapServers.toInt()
+        val sources = MagicSearch.Source.Friends.toInt() or MagicSearch.Source.LdapServers.toInt()
         val aggregation = MagicSearch.Aggregation.Friend
         searchResultsPending = true
         fastFetchJob?.cancel()
         Log.i(
-            "[Contacts] Asking Magic search for contacts matching filter [$filterValue], domain [$domain] and in sources [$filter]"
+            "[Contacts] Asking Magic search for contacts matching filter [$filterValue], domain [$domain] and in sources [$sources]"
         )
         coreContext.contactsManager.magicSearch.getContactsListAsync(
             filterValue,
             domain,
-            filter,
+            sources,
             aggregation
         )